🔧 직관리스트 폴더 구조 변경
